feat(callback): remove the `onNewValue` callback (#1)
The hook was making an opinionated choice to parse the current value with `parseFloat` before
passing it to `onNewValue`. This was adding unnecessary mental overhead and preventing the caller
from performing certain transformations on the new value. Additionally, the more idiomatic way to
handle `value` updates is to use `React.useEffect`, specifying `value` in the dependency array.
With those two notions in mind, the callback no longer serves any purpose.

BREAKING CHANGE: The `onNewValue` callback option will be ignored and no longer called. To respond
to value updates, consider using `React.useEffect` with `value` specified in the dependency array.
As the argument to `onNewValue` was previously a number and `value` is a string, your logic may
need adjusted to parse `value` if you were relying on the new value being a number in your update
handler.