Switch to npm trusted publishing (#252)
* Switch to npm trusted publishing

* Replace jsdom with happy-dom

This has nothing to do with what I was trying to do in this PR (switch
to npm trusted publishing) but tests were failing in CI claiming they
couldn't find jsdom. This fixes that. I don't know why it couldn't find
it and I couldn't care less (jsdom sucks anyway).